vulnScanner/requirements.txt
streamlit>=1.10.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-whois>=0.8.0
reportlab>=3.6.0
supabase>=0.7.0
python-dotenv>=0.20.0
dnspython>=2.3.0
responses>=0.25.0
pytest>=8.0.0
pytest-mock>=3.14.0